# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Unified observability platform for AI agents, APIs, and infrastructure.",
        "title": "Keelwave API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/v1",
    "paths": {
...
        "/ingest/agent/runs": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Opens a new agent run record with status=\"running\". Returns id + timestamp; the client must pass both back to finish the run.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Start an agent run",
                "parameters": [
                    {
                        "description": "Run start payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ingestAgentRunStartPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.ingestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/ingest/agent/runs/{runID}/finish": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Marks a run terminal and writes totals. timestamp must match the value returned at start (hypertable PK).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Finish an agent run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent run UUID",
                        "name": "runID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Run finish payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ingestAgentRunFinishPayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/projects/{projectID}/agent/runs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "List agent runs for the calling project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now - 30d",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 25, max 1000",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.AgentRun"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/projects/{projectID}/agent/runs/terminations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "Run counts grouped by termination reason",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now - 30d",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.TerminationCount"
                            }
                        }
                    }
                }
            }
        },
        "/projects/{projectID}/agent/runs/timeseries": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns per-bucket run outcome counts (total, completed, failed, loop) via TimescaleDB time_bucket. bucket is one of 1h, 6h, 1d.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "Bucketed run counts over time for the calling project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now - 30d",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "bucket size: 1h (default), 6h, 1d",
                        "name": "bucket",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.RunBucket"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/projects/{projectID}/agent/runs/{runID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "Fetch a single agent run by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Agent run UUID",
                        "name": "runID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339; if set, lookup is bounded to a 2s window for fast chunk pruning",
                        "name": "at",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.AgentRun"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/projects/{projectID}/agent/summary": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agent"
                ],
                "summary": "Run summary + duration percentiles for the calling project, current vs previous window",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now - 30d",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339, default now",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/store.RunSummary"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.ingestAgentRunFinishPayload": {
            "type": "object",
            "required": [
                "status",
                "timestamp"
            ],
            "properties": {
                "duration_ms": {
                    "type": "integer",
                    "minimum": 0
                },
                "loop_detected": {
                    "type": "boolean"
                },
                "loop_step_index": {
                    "type": "integer",
                    "minimum": 0
                },
                "output": {
                    "type": "string",
                    "maxLength": 100000
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "completed",
                        "failed"
                    ]
                },
                "termination_reason": {
                    "type": "string",
                    "enum": [
                        "clean",
                        "max_steps_reached",
                        "context_limit",
                        "error",
                        "loop_detected",
                        "timeout"
                    ]
                },
                "timestamp": {
                    "type": "string"
                },
                "total_cost_usd": {
                    "type": "number",
                    "minimum": 0
                },
                "total_steps": {
                    "type": "integer",
                    "minimum": 0
                },
                "total_tokens": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "main.ingestAgentRunStartPayload": {
            "type": "object",
            "required": [
                "agent_name"
            ],
            "properties": {
                "agent_name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "input": {
                    "type": "string",
                    "maxLength": 10000
                },
                "metadata": {
                    "type": "object"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "main.ingestResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "store.AgentRun": {
            "type": "object",
            "properties": {
                "agent_name": {
                    "type": "string"
                },
                "duration_ms": {
                    "type": "integer"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "loop_detected": {
                    "type": "boolean"
                },
                "loop_step_index": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "output": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "termination_reason": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "total_cost_usd": {
                    "type": "number"
                },
                "total_steps": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
...
        "store.RunBucket": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "completed": {
                    "type": "integer"
                },
                "failed": {
                    "type": "integer"
                },
                "loop": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "store.RunSummary": {
            "type": "object",
            "properties": {
                "avg_cost_usd": {
                    "type": "number"
                },
                "avg_tokens": {
                    "type": "number"
                },
                "completed_runs": {
                    "type": "integer"
                },
                "completion_rate": {
                    "type": "number"
                },
                "duration_p50_ms": {
                    "type": "integer"
                },
                "duration_p95_ms": {
                    "type": "integer"
                },
                "duration_p99_ms": {
                    "type": "integer"
                },
                "loop_rate": {
                    "type": "number"
                },
                "loop_runs": {
                    "type": "integer"
                },
                "total_runs": {
                    "type": "integer"
                },
                "total_steps": {
                    "type": "integer"
                },
                "total_tool_calls": {
                    "type": "integer"
                },
                "unique_agents": {
                    "type": "integer"
                },
                "unique_tools": {
                    "type": "integer"
                }
            }
        },
...
        "store.TerminationCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "termination_reason": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Format: \"Bearer kw_\u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}